[FE/#10] Refactor : menu list를 다른 컴포넌트로 분리
